+++ src/types/options/index.ts
@@ -9,3 +9,4 @@ export * from './RenderingOptions';
 export * from './LoadingIndicatorOptions';
 export * from './LoaderOptions';
 export * from './PresetOptions';
+export * from './UIOptions';